Code review: Fix AsioSendBuffer deadlock, spurious lobby notifications, OOB write, sqlite3_free_table leak, accumulating signal connections, uninitialized pointers, timestamp truncation, and data loss on parse errors

+++ src/db/serverdbinterface.h
@@ -58,7 +58,7 @@ public:
 
 	virtual void AsyncCreateGame(unsigned requestId, const std::string &gameName) = 0;
 	virtual void SetGamePlayerPlace(unsigned requestId, DB_id playerId, unsigned place) = 0;
-	virtual void SetPlayerLastGames(unsigned requestId, DB_id playerId, const std::vector<long> &last_games, const std::string &playerIp) = 0;
+	virtual void SetPlayerLastGames(unsigned requestId, DB_id playerId, const std::vector<long long> &last_games, const std::string &playerIp) = 0;
 	virtual void EndGame(unsigned requestId) = 0;
 
 	virtual void AsyncReportAvatar(unsigned requestId, unsigned replyId, DB_id reportedPlayerId, const std::string &avatarHash, const std::string &avatarType, DB_id *byPlayerId) = 0;
